refactor: 배치 첫 스텝 중간 테이블 컬럼 변경
