+++ src/main/java/com/example/masterplanbbe/application/batch/dto/FirstStepReadDTO.java
@@ -3,6 +3,7 @@ package com.example.masterplanbbe.application.batch.dto;
 import java.time.LocalDate;
 
 public record FirstStepReadDTO(
+        Long specId,
         Long examId,
         LocalDate applyEndDate,
         LocalDate examStartDate,
