+++ src/main/java/com/example/masterplanbbe/application/batch/dto/FirstStepWriteDTO.java
@@ -1,4 +1,4 @@
 package com.example.masterplanbbe.application.batch.dto;
 
-public record FirstStepWriteDTO(Long examId, Double intermediateResult) {
+public record FirstStepWriteDTO(Long specId, Long examId, Double intermediateResult) {
 }
